תיאור קצר של השינוי שביצעתי

+++ Bill3.java
@@ -7,7 +7,7 @@ public class Bill3 {
 		String name3 = args[2];
 		int all = Integer.parseInt(args[3]);
 		
-		double pay  = (double) all/3;
+		int pay  =  all/3;
 	   
 		System.out.println("Dear " + name3 + ", " + name2 + " and " + name1 + ": pay " + pay + " Shekels");
 	}
